LPS-166857 Moves to layout-impl
