+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/LayoutStructureItemImporterRegistry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.headless.delivery.dto.v1_0.structure.importer;
+package com.liferay.layout.internal.importer;
 
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
