+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/exception/DropzoneLayoutStructureItemException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.exception;
+package com.liferay.layout.internal.importer.exception;
 
 import com.liferay.portal.kernel.exception.PortalException;
