+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/structure/util/FormLayoutStructureItemImporter.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.headless.delivery.dto.v1_0.structure.importer;
+package com.liferay.layout.internal.importer.structure.util;
 
 import com.liferay.headless.delivery.dto.v1_0.ContextReference;
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
+import com.liferay.layout.internal.importer.LayoutStructureItemImporterContext;
 import com.liferay.layout.page.template.util.AlignConverter;
 import com.liferay.layout.page.template.util.ContentDisplayConverter;
 import com.liferay.layout.page.template.util.FlexWrapConverter;
