+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/structure/util/FragmentDropZoneLayoutStructureItemImporter.java
@@ -12,26 +12,40 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.headless.delivery.dto.v1_0.structure.importer;
+package com.liferay.layout.internal.importer.structure.util;
 
 import com.liferay.headless.delivery.dto.v1_0.PageElement;
+import com.liferay.layout.internal.importer.LayoutStructureItemImporterContext;
 import com.liferay.layout.util.structure.LayoutStructure;
 import com.liferay.layout.util.structure.LayoutStructureItem;
 
 import java.util.Set;
 
+import org.osgi.service.component.annotations.Component;
+
 /**
  * @author Jürgen Kappler
  */
-public interface LayoutStructureItemImporter {
+@Component(service = LayoutStructureItemImporter.class)
+public class FragmentDropZoneLayoutStructureItemImporter
+	implements LayoutStructureItemImporter {
 
+	@Override
 	public LayoutStructureItem addLayoutStructureItem(
 			LayoutStructure layoutStructure,
 			LayoutStructureItemImporterContext
 				layoutStructureItemImporterContext,
 			PageElement pageElement, Set<String> warningMessages)
-		throws Exception;
+		throws Exception {
 
-	public PageElement.Type getPageElementType();
+		return layoutStructure.addFragmentDropZoneLayoutStructureItem(
+			layoutStructureItemImporterContext.getParentItemId(),
+			layoutStructureItemImporterContext.getPosition());
+	}
+
+	@Override
+	public PageElement.Type getPageElementType() {
+		return PageElement.Type.FRAGMENT_DROP_ZONE;
+	}
 
 }
